feature: 业务集形态改造（平台内置转对接CMDB） #622
format
