+++ src/backend/commons/cmdb-sdk/src/main/java/com/tencent/bk/job/common/cc/util/TopologyUtil.java
@@ -31,7 +31,6 @@ import lombok.extern.slf4j.Slf4j;
 
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
