+++ src/backend/commons/common/src/main/java/com/tencent/bk/job/common/app/BasicAppTransferService.java
@@ -1,0 +1,63 @@
+/*
+ * Tencent is pleased to support the open source community by making BK-JOB蓝鲸智云作业平台 available.
+ *
+ * Copyright (C) 2021 THL A29 Limited, a Tencent company.  All rights reserved.
+ *
+ * BK-JOB蓝鲸智云作业平台 is licensed under the MIT License.
+ *
+ * License for BK-JOB蓝鲸智云作业平台:
+ * --------------------------------------------------------------------
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
+ * documentation files (the "Software"), to deal in the Software without restriction, including without limitation
+ * the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and
+ * to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or substantial portions of
+ * the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO
+ * THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF
+ * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+ * IN THE SOFTWARE.
+ */
+
+package com.tencent.bk.job.common.app;
+
+import org.springframework.stereotype.Service;
+
+import java.util.Collection;
+import java.util.Map;
+
+/**
+ * 业务类型转换基础类
+ */
+@Service
+public abstract class BasicAppTransferService implements AppTransferService {
+    @Override
+    public abstract Long getAppIdByScope(ResourceScope resourceScope);
+
+    @Override
+    public void fillResourceScope(ResourceScope resourceScope) {
+        if (resourceScope.isFull()) {
+            return;
+        }
+        if (resourceScope.hasAppId()) {
+            ResourceScope scope = getScopeByAppId(resourceScope.getAppId());
+            resourceScope.setType(scope.getType());
+            resourceScope.setId(scope.getId());
+            return;
+        }
+        Long appId = getAppIdByScope(resourceScope);
+        resourceScope.setAppId(appId);
+    }
+
+    @Override
+    public abstract ResourceScope getScopeByAppId(Long appId);
+
+    @Override
+    public abstract Map<Long, ResourceScope> getScopeByAppIds(Collection<Long> appIds);
+
+    @Override
+    public abstract ResourceScope getResourceScope(Long appId, String scopeType, String scopeId);
+}
