+++ src/backend/job-manage/service-job-manage/src/main/java/com/tencent/bk/job/manage/auth/impl/AccountAuthServiceImpl.java
@@ -24,7 +24,6 @@
 
 package com.tencent.bk.job.manage.auth.impl;
 
-import com.tencent.bk.job.common.app.AppTransferService;
 import com.tencent.bk.job.common.app.ResourceScope;
 import com.tencent.bk.job.common.iam.constant.ActionId;
 import com.tencent.bk.job.common.iam.constant.ResourceTypeEnum;
